Debugging opengl.cpp on Windows.

+++ Code/include/mesh.hpp
@@ -31,7 +31,7 @@ private:
   size_t                                 k;                                                         ///< Index.
   size_t                                 n;                                                         ///< Index.
   size_t                                 dim;                                                       ///< Entity dimension.
-  size_t                                 tag;                                                       ///< Entity tag.
+  int                                    tag;                                                       ///< Entity tag.
   std::vector<std::pair<int, int> >      entities;                                                  ///< Entity array.
   std::vector<std::size_t>               nodeTags;                                                  ///< Node tags array...
   std::vector<double>                    nodeCoords;                                                ///< Node coordinates array...
